requirements.txt
numpy
pillow
opencv-python
ffmpeg-python
torch
torchvision
tqdm
matplotlib